Move hal_bb_gpio sources to own subdir & wrap lines
This should be a no-op:
- Move hal_bb_gpio sources to `hal/drivers/hal_bb_gpio`
- Set indentation to 4 spaces and re-wrap long lines
